product, auth and cart added

+++ src/main/java/com/groupthree/shopsphere/repository/UserRepository.java
@@ -3,6 +3,7 @@ package com.groupthree.shopsphere.repository;
 import com.groupthree.shopsphere.models.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User,Integer> {
+public interface UserRepository extends CrudRepository<User,Long> {
     User findByEmail(String email);
+
 }
